Reactチュートリアル タイムトラベル機能の追加.Stateのリフトアップ、再び
※着手の履歴保存ではコード変更なし。
