+++ React-tutorial/my-app/src/Board.tsx
@@ -1,9 +1,11 @@
 import React from 'react';
 import './index.css';
 import { Square } from './Square';
-import { calculateWinner } from './calculateWinner';
 
-interface BoardProps {}
+interface BoardProps {
+  squares: Array<string>;
+  onClick: (i: number) => void;
+}
 
 interface BoardState {
   squares: Array<string>;
@@ -11,47 +13,18 @@ interface BoardState {
 }
 
 export class Board extends React.Component<BoardProps, BoardState> {
-  public constructor(props: BoardProps) {
-    super(props);
-    this.state = {
-      squares: Array(9).fill(''),
-      xIsNext: true,
-    };
-  }
-
-  private handleClick(i: number) {
-    const squares = this.state.squares.slice();
-    if (calculateWinner(squares) || squares[i]) {
-      return;
-    }
-    squares[i] = this.state.xIsNext ? 'X' : 'O';
-    this.setState({
-      squares: squares,
-      xIsNext: !this.state.xIsNext,
-    });
-  }
-
   private renderSquare(i: number) {
     return (
       <Square
-        value={this.state.squares[i]}
-        onClick={() => this.handleClick(i)}
+        value={this.props.squares[i]}
+        onClick={() => this.props.onClick(i)}
       />
     );
   }
 
   public render() {
-    const winner = calculateWinner(this.state.squares);
-    let status: string;
-    if (winner) {
-      status = 'Winner: ' + winner;
-    } else {
-      status = 'Next Player: ' + (this.state.xIsNext ? 'X' : 'O');
-    }
-
     return (
       <div>
-        <div className="status">{status}</div>
         <div className="board-row">
           {this.renderSquare(0)}
           {this.renderSquare(1)}
