+++ React-tutorial/my-app/src/index.tsx
@@ -2,28 +2,77 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { Board } from './Board';
 import './index.css';
-  
-  class Game extends React.Component {
-    public render() {
-      return (
-        <div className="game">
-          <div className="game-board">
-            <Board />
-          </div>
-          <div className="game-info">
-            <div>{/* status */}</div>
-            <ol>{/* TODO */}</ol>
-          </div>
-        </div>
-      );
-    }
+import { calculateWinner } from './calculateWinner';
+
+interface Squares {
+  squares: Array<string>;
+}
+
+interface GameState {
+  history: Squares[];
+  xIsNext: boolean;
+}
+
+class Game extends React.Component<{}, GameState> {
+  public constructor(props: {}) {
+    super(props);
+    this.state = {
+      history: [{
+        squares: Array(9).fill(''),
+      }],
+      xIsNext: true,
+    };
   }
-  
-  // ========================================
-  
-  ReactDOM.render(
-    <Game />,
-    document.getElementById('root')
-  );
-  
+
+  private handleClick(i: number) {
+    const history = this.state.history;
+    const current = history[history.length - 1];
+    const squares = current.squares.slice();
+    if (calculateWinner(squares) || squares[i]) {
+      return;
+    }
+    squares[i] = this.state.xIsNext ? 'X' : 'O';
+    this.setState({
+      history: history.concat([{
+        squares: squares,
+      }]),
+      xIsNext: !this.state.xIsNext,
+    });
+  }
+
+  public render() {
+    const history = this.state.history;
+    const current = history[history.length - 1];
+    const winner = calculateWinner(current.squares);
+    let status: string;
+    if (winner) {
+      status = 'Winner: ' + winner;
+    } else {
+      status = 'Next player: ' + (this.state.xIsNext ? 'X' : 'O');
+    }
+
+    return (
+      <div className="game">
+        <div className="game-board">
+          <Board
+            squares={current.squares}
+            onClick={(i) => this.handleClick(i)}
+          />
+        </div>
+        <div className="game-info">
+          <div>{status}</div>
+          <ol>{/* TODO */}</ol>
+        </div>
+      </div>
+    );
+  }
+}
+
+// ========================================
+
+ReactDOM.render(
+  <Game />,
+  document.getElementById('root')
+);
+
